requirements.txt
discord.py
promptpay
qrcode[pil]